+++ docs/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @inkeep/open-knowledge-docs
 
+## 0.0.210
+
+### Patch Changes
+
+- @inkeep/open-knowledge-core@0.76.0
+
+## 0.0.209
+
+### Patch Changes
+
+- @inkeep/open-knowledge-core@0.75.1
+
 ## 0.0.208
 
 ### Patch Changes